--- components/atoms/List/List/index.ts
@@ -1,2 +1,0 @@
-export { List } from './List'
-export { listStyle } from './style'
