+++ components/atoms/List/ListItem/index.ts
@@ -1,0 +1,2 @@
+export { ListItem } from './ListItem'
+export { listItemStyle } from './style'
