+++ components/atoms/List/ListItem/style.ts
@@ -1,6 +1,6 @@
 import { css } from '@emotion/react'
 
-export const listStyle = css`
+export const listItemStyle = css`
   line-height: 1.2;
   padding: 0.5rem 0;
 `
