+++ components/atoms/List/index.ts
@@ -1,3 +1,3 @@
-export * from './List'
+export * from './ListItem'
 export * from './OrderedList'
 export * from './UnorderedList'
