+++ components/atoms/Typography/Label/index.ts
