+++ components/atoms/Typography/Label/style.ts
@@ -1,0 +1,7 @@
+import { css } from '@emotion/react'
+import { theme } from '@/theme/theme'
+
+export const style = css`
+  font-size: ${theme.fontSizes.S};
+  margin: 1rem 0;
+`
