+++ components/atoms/Typography/index.ts
@@ -2,5 +2,6 @@ export * from './H1'
 export * from './H2'
 export * from './H3'
 export * from './H4'
+export * from './Label'
 export * from './Paragraph'
 export * from './Blockquote'
