+++ components/molecules/Tag/index.tsx
@@ -1,10 +1,11 @@
 import React from 'react'
 import { Anchor } from '@/components/atoms/Anchor'
-import { Label } from '@/components/atoms/Label'
+import { Label } from '@/components/atoms/Typography'
+import { style } from './style'
 
 const Tag: React.FC<{ href?: string; text: string }> = ({ href, text }) => {
   return (
-    <Label>
+    <Label css={style}>
       <Anchor href={href}>{text}</Anchor>
     </Label>
   )
