+++ components/molecules/Tag/style.ts
